+++ src/Components/ReviewLayout/WatchList/WatchReviewList.jsx
@@ -2,7 +2,9 @@ import { MdOutlineDelete } from "react-icons/md";
 import Swal from "sweetalert2";
 
 const WatchReviewList = ({ review, reviews, setReviews }) => {
-  const handleDelete = (id) => {
+  const { _id, name, email, image, title, description, rating, year, genres } =
+    review;
+  const handleDelete = (_id) => {
     Swal.fire({
       title: "Are you sure?",
       text: "You won't be able to revert this!",
@@ -11,25 +13,29 @@ const WatchReviewList = ({ review, reviews, setReviews }) => {
       confirmButtonColor: "#3085d6",
       cancelButtonColor: "#d33",
       confirmButtonText: "Yes, delete it!",
-    });
-    fetch(`http://localhost:5000/watchList/${review._id}`, {
-      method: "DELETE",
-    })
-      .then((res) => res.json())
-      .then((data) => {
-        // console.log(data);
-        if (data.deletedCount < 0) {
-          Swal.fire({
-            title: "Deleted!",
-            text: "Your WatchList has been deleted.",
-            icon: "success",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        fetch(
+          `https://assignment-10-server-beta-steel.vercel.app/watchList/${review._id}`,
+          {
+            method: "DELETE",
+          }
+        )
+          .then((res) => res.json())
+          .then((data) => {
+            // console.log(data);
+            if (data.deletedCount > 0) {
+              Swal.fire({
+                title: "Deleted!",
+                text: "Your WatchList has been deleted.",
+                icon: "success",
+              });
+              const deleteWatchReview = reviews.filter((r) => r._id !== _id);
+              setReviews(deleteWatchReview);
+            }
           });
-          const deleteWatchReview = reviews.filter(
-            (review) => review._id !== id
-          );
-          setReviews(deleteWatchReview);
-        }
-      });
+      }
+    });
   };
 
   return (
@@ -40,31 +46,31 @@ const WatchReviewList = ({ review, reviews, setReviews }) => {
       >
         <input type="checkbox" />
         <div className="collapse-title text-xl font-medium text-white ">
-          <p>watchList: {review.name}</p>
+          <p>watchList: {name}</p>
         </div>
         <div className="collapse-content text-white lg:flex gap-10 ">
           <div className="card card-side flex justify-center shadow-xl lg:w-4/12">
             <figure className="">
-              <img src={review.image} alt="Movie" className="rounded-xl" />
+              <img src={image} alt="Movie" className="rounded-xl" />
             </figure>
           </div>
 
           <div className="lg:w-full">
             <div className=" text-center lg:text-start ">
               <span className="badge badge-ghost bg-slate-800 text-white py-4 shadow-2xl mr-3">
-                {review.title}
+                {title}
               </span>
               <span className="badge badge-ghost bg-slate-800 text-white py-4 px-4 mr-3">
-                {review.genres}
+                {genres}
               </span>
               <span className="badge badge-ghost bg-slate-800 text-white py-4 px-4 mr-3">
-                {review.year}
+                {year}
               </span>
-              <p className="mt-5 text-start ">{review.description}</p>
+              <p className="mt-5 text-start ">{description}</p>
             </div>
             <div className="lg:mt-52 flex justify-end mr-4">
               <button
-                onClick={() => handleDelete(review._id)}
+                onClick={() => handleDelete(_id)}
                 className="flex items-center gap-1 text-gray-300"
               >
                 <p>delete game</p>
